4.15 base uploaded

+++ include/linux/mm_types_task.h
@@ -32,14 +32,10 @@
 #define VMACACHE_MASK (VMACACHE_SIZE - 1)
 
 struct vmacache {
-	u64 seqnum;
+	u32 seqnum;
 	struct vm_area_struct *vmas[VMACACHE_SIZE];
 };
 
-/*
- * When updating this, please also update struct resident_page_types[] in
- * kernel/fork.c
- */
 enum {
 	MM_FILEPAGES,	/* Resident file mapping pages */
 	MM_ANONPAGES,	/* Resident anonymous pages */
